Abins 2D: Add MERLIN, refactor 2D instruments
Introduce a DirectInstrument class which handles e_init setting and
kinematic constraints; this reduces redundancy and keeps
instrument-specific classes more focused.

MERLIN is just another Pychop instrument, so very straightforward to
add. Some of these Pychop parameters are consistent actually, maybe
that redundancy can also be cut...
