+++ scripts/abins/parameters.py
@@ -39,8 +39,8 @@ instruments = {
         'angle_range': (5, 140),
         'chopper_frequency_default': 400,
         'chopper_allowed_frequencies': list(range(50, 601, 50)),
-        'max_wavenumber': 1613,  # maximum wavenumber in cm^-1 taken into account while creating workspaces (exclusive)
-        'min_wavenumber': 0.0,  # minimal wavenumber in cm^-1 taken into account while creating workspaces (exclusive)
+        'max_wavenumber': 150 * MILLI_EV_TO_WAVENUMBER,
+        'min_wavenumber': 0.0,
         # Resolution function fitted to incident energy and energy transfer:
         # sigma = polyval(abs_meV, ɛ) + polyval(ei_dependence, E_i) + polyval(ei_energy_product, E_i × ɛ)
         # (Here a quartic polynomial in ɛ, plus quadratic on Ei and cubic on ɛ×Ei)
@@ -73,8 +73,18 @@ instruments = {
         'e_init': 400 * MILLI_EV_TO_WAVENUMBER,
         'settings_default': 'A',
         'settings': {'A': {'chopper': 'A'},
-                     'R': {'chopper': 'A'},
-                     'G': {'chopper': 'S'},
+                     'G': {'chopper': 'G'},
+                     'R': {'chopper': 'R'},
+                     'S': {'chopper': 'S'},},
+        'chopper_frequency_default': 400,
+        'chopper_allowed_frequencies': list(range(50, 601, 50))
+        },
+    'MERLIN': {
+        'resolution': 'pychop',
+        'q_size': 100,
+        'e_init': 400 * MILLI_EV_TO_WAVENUMBER,
+        'settings_default': 'G',
+        'settings': {'G': {'chopper': 'G'},
                      'S': {'chopper': 'S'},},
         'chopper_frequency_default': 400,
         'chopper_allowed_frequencies': list(range(50, 601, 50))
